*Improved the responsivness of arrowing around an internet explorer virtual buffer document by keeping info about each field, such as its typeString (link, list, list item, heading etc) and also keeping a reference to its dom node, rather than having to ask for it by its ID every time. The ID also knows about two optional functions that can return state text (such as checked) and description text (such as "with 5 items"). These functions are optional so this means that only certain IDs will have to access their dom node for info. *When a new page is loaded in an internet explorer frame, the virtual buffer is reloaded. It would be nice to just refresh that part of the buffer which has changed, but onreadystatechange event for some odd reason does not provide an srcElement or fromElement or toElement or relatedTarget property, so it is impossible to see exactly who the event was fired for. All we can do is check the main document's readyState value, and if it is "complete" then we reload. *start and end frame messages in internet explorer virtual buffer objects are no longer phisically printed to the virtual buffer, but frames now have enter and exit messages as you move in and out of them the same as lists do. *It is now possible to update any part of the virtual buffer (if of course you have a good enough internet explorer event that can tell you exactly what dom node it is for). Because of event problems, random updates by javascript etc still doesn't happen, but it is possible if events can be fixed up.
Assuming that most of this doesn't break in IE 7, (which I am still unable to test) this should make all general websites (including forms and frames) accessible enough. But some banking and shopping sites still will be partly inaccessible because of popup windows and dynamic content.

+++ source/NVDAObjects/MSAA.py
@@ -860,7 +860,7 @@ class NVDAObject_progressBar(NVDAObject_MSAA):
 	def event_valueChange(self):
 		if conf["presentation"]["beepOnProgressBarUpdates"]:
 			baseFreq=440
-			winsound.Beep(baseFreq*(1+(float(self.value[:-1])/100.0)),100)
+			winsound.Beep(int(baseFreq*(1+(float(self.value[:-1])/100.0))),100)
 		super(NVDAObject_progressBar,self).event_valueChange()
 
 ###class mappings
